editor: Make AABB drawing generic.
