--- VisualPinball.Unity/VisualPinball.Unity.Editor/EditorTool/PhysicsTool.cs
@@ -1,10 +1,0 @@
-using UnityEditor.EditorTools;
-
-namespace VisualPinball.Unity.Editor
-{
-	[EditorTool("Platform Tool")]
-	public class PhysicsTool : EditorTool
-	{
-
-	}
-}
