--- VisualPinball.Unity/VisualPinball.Unity.Editor/EditorTool/PhysicsTool.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9f0b07d9a4fcd364da833efb8443c110
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
