+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Flipper/FlipperAuthoring.cs
@@ -24,7 +24,6 @@ namespace VisualPinball.Unity
 		protected override Flipper GetItem() => new Flipper(data);
 
 		private static readonly Color EndAngleMeshColor = new Color32(0, 255, 248, 10);
-		private static readonly Color AabbColor = new Color32(255, 0, 252, 255);
 
 		private void OnDestroy()
 		{
@@ -74,9 +73,9 @@ namespace VisualPinball.Unity
 			data.Height = scale.z;
 		}
 
-		private void OnDrawGizmosSelected()
+		protected override void OnDrawGizmosSelected()
 		{
-			var ltw = transform.GetComponentInParent<TableAuthoring>().gameObject.transform.localToWorldMatrix;
+			base.OnDrawGizmosSelected();
 
 			// draw end position mesh
 			var mfs = GetComponentsInChildren<MeshFilter>();
@@ -91,18 +90,6 @@ namespace VisualPinball.Unity
 				var r = math.mul(baseRotation, quaternion.EulerXYZ(0, 0, data.EndAngle));
 				Gizmos.DrawWireMesh(mf.sharedMesh, t.position, r, t.lossyScale);
 			}
-
-			Item.Init(Table);
-			var hitObj = Item.GetHitShapes()[0];
-			hitObj.CalcHitBBox();
-			var aabb = hitObj.HitBBox;
-
-			const float s = 0.1f;
-			var center = new Vector3(data.Center.X, data.Center.Y, 0f);
-			Gizmos.color = AabbColor;
-			DrawAabb(ltw, aabb);
-			//Gizmos.DrawCube(ltw.MultiplyPoint(center), new Vector3(s, s, s));
-			//Gizmos.DrawCube(ltw * aabb.Center.ToUnityVector3(), new Vector3(s, s, s));
 		}
 
 		private FlipperStaticData GetMaterialData()
@@ -145,35 +132,6 @@ namespace VisualPinball.Unity
 				EndRadius = endRadius,
 				FlipperRadius = flipperRadius
 			};
-		}
-
-		protected void DrawAabb(Matrix4x4 ltw, Rect3D aabb)
-		{
-			var p00 = ltw.MultiplyPoint(new Vector3( aabb.Left, aabb.Top, aabb.ZHigh));
-			var p01 = ltw.MultiplyPoint(new Vector3(aabb.Left, aabb.Bottom, aabb.ZHigh));
-			var p02 = ltw.MultiplyPoint(new Vector3(aabb.Right, aabb.Bottom, aabb.ZHigh));
-			var p03 = ltw.MultiplyPoint(new Vector3(aabb.Right, aabb.Top, aabb.ZHigh));
-
-			var p10 = ltw.MultiplyPoint(new Vector3( aabb.Left, aabb.Top, aabb.ZLow));
-			var p11 = ltw.MultiplyPoint(new Vector3(aabb.Left, aabb.Bottom, aabb.ZLow));
-			var p12 = ltw.MultiplyPoint(new Vector3(aabb.Right, aabb.Bottom, aabb.ZLow));
-			var p13 = ltw.MultiplyPoint(new Vector3(aabb.Right, aabb.Top, aabb.ZLow));
-
-			Gizmos.color = AabbColor;
-			Gizmos.DrawLine(p00, p01);
-			Gizmos.DrawLine(p01, p02);
-			Gizmos.DrawLine(p02, p03);
-			Gizmos.DrawLine(p03, p00);
-
-			Gizmos.DrawLine(p10, p11);
-			Gizmos.DrawLine(p11, p12);
-			Gizmos.DrawLine(p12, p13);
-			Gizmos.DrawLine(p13, p10);
-
-			Gizmos.DrawLine(p00, p10);
-			Gizmos.DrawLine(p01, p11);
-			Gizmos.DrawLine(p02, p12);
-			Gizmos.DrawLine(p03, p13);
 		}
 
 		private FlipperMovementData GetMovementData(FlipperStaticData d)
